step03 ing

+++ src/main/java/org/homework/TodoController.java
@@ -1,8 +1,21 @@
 package org.homework;
 
+import java.text.ParseException;
+import java.text.SimpleDateFormat;
+import java.util.Comparator;
+import java.util.Date;
+import java.util.NoSuchElementException;
+import java.util.Optional;
+import java.util.stream.Stream;
+
 public class TodoController {
-    private TodoService todoService;
-    public void run() {
+    private final TodoService todoService;
+
+    public TodoController() {
+        this.todoService = new TodoService();
+    }
+
+    public void run(){
         while (true) {
             switch (choiceMenu()) {
                 case ADD:
@@ -12,12 +25,15 @@ public class TodoController {
                     delete();
                     break;
                 case SELECT:
-                    select();
+                    selectAll();
+//                    select();
                     break;
                 case EXIT:
                     exit();
                     return;
                 case NONE:
+                    error();
+                    break;
             }
         }
     }
@@ -25,19 +41,49 @@ public class TodoController {
         OutputView.outPutChoiceMenu();
         return InputView.inputMenu();
     }
-    private void save() {
-        todoService.save(InputView.inputContent());
+    private void save(){
+        String content = InputView.inputString();
+        Date date = inputDate();
+        Todo todo = new Todo(content,date);
+        OutputView.outPutAddTodo(todoService.save(todo));
     }
     private void delete() {
         int num = InputView.inputNum();
-        todoService.delete(num);
-        OutputView.outPutDeleteTodo(num);
-    }
-    private void select() {
-        OutputView.outPutSelectTodo(todoService.findById(InputView.inputNum()));
+        if (todoService.existsById(num)) {
+            todoService.delete(num);
+        } else {
+            OutputView.outPutTodoIfNull();
+        }
     }
 
+    private void select() {
+        Optional<Todo> findByIdTodo = todoService.findById(InputView.inputNum());
+        if (findByIdTodo.isPresent()) {
+            OutputView.outPutSelectTodo(findByIdTodo.get());
+        } else {
+            OutputView.outPutTodoIfNull();
+        }
+
+    }
+    private void selectAll() {
+        OutputView.outPutTodoAll(todoService.selectAll());
+    }
     private void exit() {
         OutputView.outPutFinishProgram();
     }
+    private void error() {
+        OutputView.outPutError();
+    }
+
+    private Date inputDate() {
+        OutputView.outPutAddEndDate();
+        String endDateStr = InputView.inputString();
+        SimpleDateFormat format = new SimpleDateFormat("yyyyMMdd");
+        try {
+            return format.parse(endDateStr);
+        } catch (ParseException e) {
+            return null;
+        }
+    }
+
 }
